test: migrate notifications smoke Detox tests to Appium (MMQA-1996)
Move enable-notifications-after-onboarding and notification utils into
smoke-appium, update consumers, and make NotificationMenuView scroll
cross-framework.
