+++ tests/component-view/presets/notifications.ts
@@ -12,8 +12,8 @@ import { AvatarAccountType } from '../../../app/component-library/components/Ava
  * but seeded directly into Redux instead of fetched via mockttp.
  *
  * Smoke E2E refs:
- * - tests/smoke/notifications/enable-notifications-after-onboarding.spec.ts
- * - tests/smoke/notifications/utils/mocks.ts
+ * - tests/smoke-appium/notifications/enable-notifications-after-onboarding.spec.ts
+ * - tests/smoke-appium/notifications/utils/mocks.ts
  */
 export const MOCK_NOTIFICATIONS = [
   ...MOCK_FEATURE_ANNOUNCEMENT_NOTIFICATIONS,
