+++ tests/smoke-appium/notifications/utils/constants.ts
