+++ tests/smoke-appium/notifications/utils/helpers.ts
