+++ tests/smoke-appium/notifications/utils/mock-notification-trigger-server.ts
