+++ tests/smoke-appium/notifications/utils/mocks.ts
@@ -28,7 +28,7 @@ import {
 } from '@metamask/notification-services-controller/push-services/mocks';
 import { getDecodedProxiedURL } from './helpers';
 import { MockttpNotificationTriggerServer } from './mock-notification-trigger-server';
-import { mockAuthServices } from '../../identity/utils/mocks';
+import { mockAuthServices } from '../../../smoke/identity/utils/mocks';
 import { setupMockRequest } from '../../../api-mocking/helpers/mockHelpers';
 import { createLogger } from '../../../framework/logger';
