+++ tests/smoke-appium/trending/trending-feed.spec.ts
@@ -12,7 +12,7 @@ import {
   RWA_STOCK_ASSET_ID,
 } from '../../api-mocking/mock-responses/trending-api-mocks.js';
 import { setupMockEvents } from '../../api-mocking/helpers/mockHelpers.js';
-import { getDecodedProxiedURL } from '../../smoke/notifications/utils/helpers.js';
+import { getDecodedProxiedURL } from '../notifications/utils/helpers.js';
 import {
   remoteFeatureFlagTrendingTokensEnabled,
   remoteFeatureFlagPredictEnabled,
