+++ tests/smoke/snaps/test-snap-name-lookup.spec.ts
@@ -12,7 +12,7 @@ import BrowserView from '../../page-objects/Browser/BrowserView';
 import TransactionConfirmView from '../../page-objects/Send/TransactionConfirmView';
 import TokenOverview from '../../page-objects/wallet/TokenOverview';
 import NetworkListModal from '../../page-objects/Network/NetworkListModal';
-import { getDecodedProxiedURL } from '../notifications/utils/helpers';
+import { getDecodedProxiedURL } from '../../smoke-appium/notifications/utils/helpers';
 
 jest.setTimeout(150_000);
